nautilus_tardis::machine

Module message

Structs

BarMsg

Trades data in aggregated form, known as OHLC, candlesticks, klines etc. Not only most common time based aggregation is supported, but volume and tick count based as well. Bars are computed from tick-by-tick raw trade data, if in given interval no trades happened, there is no bar produced.

BookChangeMsg

Represents a Tardis WebSocket message for book changes.

BookLevel

Represents a single level in the order book (bid or ask).

BookSnapshotMsg

Represents a Tardis WebSocket message for book snapshots.

DerivativeTickerMsg

Derivative instrument ticker info sourced from real-time ticker & instrument channels.

DisconnectMsg

Message that marks events when real-time WebSocket connection that was used to collect the historical data got disconnected.

TradeMsg

Represents a Tardis WebSocket message for trades.

Enums

WsMessage

A Tardis Machine Server message type.