nautilus_tardis::machine

Module parse

Functions

parse_bar_msg

parse_book_change_msg_as_deltas

Parse a book change message into order book deltas, returning an error if timestamps invalid. Parse a book change message into order book deltas.

parse_book_level

Parse a single book level into an order book delta.

parse_book_msg_as_deltas

Parse raw book levels into order book deltas, returning error for invalid timestamps. Parse raw book levels into order book deltas.

parse_book_snapshot_msg_as_deltas

Parse a book snapshot message into order book deltas, returning an error if timestamps invalid. Parse a book snapshot message into order book deltas.

parse_book_snapshot_msg_as_quote

Parse a book snapshot message into a quote tick, returning an error on invalid data. Parse a book snapshot message into a quote tick.

parse_tardis_ws_message

parse_trade_msg

Parse a trade message into a trade tick, returning an error on invalid data. Parse a trade message into a trade tick.