nautilus_tardis::machine::parse

Function parse_book_change_msg_as_deltas

pub fn parse_book_change_msg_as_deltas(
    msg: BookChangeMsg,
    price_precision: u8,
    size_precision: u8,
    instrument_id: InstrumentId,
) -> Result<OrderBookDeltas_API>

Parse a book change message into order book deltas, returning an error if timestamps invalid. Parse a book change message into order book deltas.

Errors

Returns an error if timestamp fields cannot be converted to nanoseconds.