nautilus_tardis

Module parse

Functions

bar_spec_to_tardis_trade_bar_string

Converts a Posei BarSpecification to the Tardis trade bar string convention.

deserialize_trade_id

Deserialize a trade ID or generate a new UUID if empty.

deserialize_uppercase

Deserialize a string and convert to uppercase Ustr.

normalize_amount

Normalizes the given amount by truncating it to the specified decimal precision.

normalize_instrument_id

Parses a Posei instrument ID with a normalized symbol from the given Tardis exchange and symbol values.

normalize_symbol_str

parse_aggressor_side

Parses a Posei aggressor side from the given Tardis string value.

parse_bar_spec

Parses a Posei bar specification from the given Tardis string value.

parse_book_action

Parses a Posei book action inferred from the given Tardis values.

parse_instrument_id

Parses a Posei instrument ID from the given Tardis exchange and symbol values.

parse_option_kind

Parses a Posei option kind from the given Tardis enum value.

parse_order_side

Parses a Posei order side from the given Tardis string value.

parse_price

Parses a Posei price from the given value.

parse_timestamp

Parses a UNIX nanoseconds timestamp from the given Tardis microseconds value_us.